Branch and bound method update.
Tool is ready for use, able to draw gantt diamgrams of the solution, run simulations to compare scores between preemption score, EDD score and B&B score.

Results are acceptable. For instance, heuristic score is always better or equal to B&B score and B&B score is always better or equal to the EDD score. Which is expected.